Issue #985 Move useFeedback hook
